pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.2",
    "fastapi>=0.128.0",
    "google-cloud-storage>=3.7.0",
    "httpx>=0.28.1",
    "numpy>=2.4.0",
    "psycopg2-binary>=2.9.11",
    "python-dateutil>=2.9.0.post0",
    "sift-stack-py>=0.9.6",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
